requirements.txt
PyQt5>=5.15.0
matplotlib>=3.3.0
reportlab>=3.6.0
pyserial>=3.5
